+++ examples/ephemeral/main.tf
@@ -38,7 +38,7 @@ module "runners" {
   runners_lambda_zip                = "../../lambda_output/runners.zip"
 
   enable_organization_runners = true
-  runner_extra_labels         = "default,example"
+  runner_extra_labels         = ["default", "example"]
 
   # enable access to the runners via SSM
   enable_ssm_on_runners = true
@@ -78,11 +78,11 @@ module "runners" {
 
   # configure your pre-built AMI
   # enable_userdata = false
-  # ami_filter      = { name = ["github-runner-amzn2-x86_64-*"], state = ["available"] }
+  # ami_filter      = { name = ["github-runner-al2023-x86_64-*"], state = ["available"] }
   # ami_owners      = [data.aws_caller_identity.current.account_id]
 
   # or use the default AMI
-  enable_userdata = true
+  # enable_userdata = true
 
   # Enable debug logging for the lambda functions
   # log_level = "debug"
@@ -95,7 +95,7 @@ module "runners" {
   # }
 }
 
-module "webhook-github-app" {
+module "webhook_github_app" {
   source     = "../../modules/webhook-github-app"
   depends_on = [module.runners]